+++ apps/client/app/(navbar)/ProfileButton.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { useSession } from "../../src/client/auth/useSession";
-import Avatar from "../../src/home/Avatar";
 import { useProfileByAddress } from "../../src/play/hooks/useProfileByAddress";
+import Avatar from "../../src/ui/Avatar";
 import { DropdownContent, DropdownMenu, DropdownTrigger } from "../../src/ui/DropdownMenu";
 import ProfileMenu from "./ProfileMenu";
