+++ apps/client/app/(navbar)/RainbowkitWrapper.tsx
@@ -2,15 +2,22 @@
 
 import "@rainbow-me/rainbowkit/styles.css";
 
-import { RainbowKitProvider } from "@rainbow-me/rainbowkit";
+import { lightTheme, RainbowKitProvider } from "@rainbow-me/rainbowkit";
 import {
   GetSiweMessageOptions,
   RainbowKitSiweNextAuthProvider,
 } from "@rainbow-me/rainbowkit-siwe-next-auth";
 import { WagmiConfig } from "wagmi";
 
-import { theme } from "../../src/client/rainbow";
 import { chains, wagmiClient } from "../../src/client/wagmi";
+
+const theme = lightTheme({
+  accentColor: "#191919",
+  accentColorForeground: "#ffffff",
+  fontStack: "system",
+  borderRadius: "large",
+  overlayBlur: "small",
+});
 
 const getSiweMessageOptions: GetSiweMessageOptions = () => ({
   statement: "🔌 Sign in to the Wired",
