+++ apps/client/app/(navbar)/user/[id]/page.tsx
@@ -4,10 +4,10 @@ import { notFound } from "next/navigation";
 import { Suspense } from "react";
 import { z } from "zod";
 
-import Avatar from "../../../../src/home/Avatar";
 import { fetchProfile } from "../../../../src/server/helpers/fetchProfile";
 import { fetchProfileFromAddress } from "../../../../src/server/helpers/fetchProfileFromAddress";
 import { getServerSession } from "../../../../src/server/helpers/getServerSession";
+import Avatar from "../../../../src/ui/Avatar";
 import Card from "../../../../src/ui/Card";
 import { toHex } from "../../../../src/utils/toHex";
 import Spaces from "./Spaces";
