+++ apps/client/app/api/projects/[id]/assets/s3.ts
@@ -2,7 +2,7 @@ import { GetObjectCommand, PutObjectCommand } from "@aws-sdk/client-s3";
 import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
 
 import { env } from "../../../../../src/env/server.mjs";
-import { s3Client } from "../../../../../src/server/client";
+import { s3Client } from "../../../../../src/server/s3";
 import { S3Path } from "../../../../../src/utils/s3Paths";
 
 const expiresIn = 600; // 10 minutes
