+++ apps/client/app/api/publications/files.ts
@@ -2,7 +2,7 @@ import { DeleteObjectsCommand, GetObjectCommand, PutObjectCommand } from "@aws-s
 import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
 
 import { env } from "../../../src/env/server.mjs";
-import { s3Client } from "../../../src/server/client";
+import { s3Client } from "../../../src/server/s3";
 import { S3Path } from "../../../src/utils/s3Paths";
 
 export const expiresIn = 600; // 10 minutes
