--- apps/client/app/api/spaces/[id]/player-count/helper.ts
@@ -1,5 +1,0 @@
-export async function getPlayerCount(id: number) {
-  const response = await fetch(`/api/spaces/${id}/player-count`, { method: "GET" });
-  const playerCount = (await response.json()) as number;
-  return playerCount;
-}
