--- apps/client/app/api/spaces/[id]/player-count/route.ts
@@ -1,14 +1,0 @@
-import { NextRequest, NextResponse } from "next/server";
-
-import { fetchPlayerCount } from "../../../../../src/server/helpers/fetchPlayerCount";
-import { Params, paramsSchema } from "../types";
-
-// Get player count
-export async function GET(request: NextRequest, { params }: Params) {
-  const { id } = paramsSchema.parse(params);
-  const spaceId = parseInt(id);
-
-  const playerCount = await fetchPlayerCount(spaceId);
-
-  return NextResponse.json(playerCount);
-}
