--- apps/client/src/constants.ts
@@ -1,12 +1,0 @@
-export const DISCORD_URL = "https://discord.gg/cazUfCCgHJ";
-export const TWITTER_URL = "https://twitter.com/wired_xr";
-export const GITHUB_URL = "https://github.com/wired-labs/wired";
-
-export const CACHED_PATHS = [
-  "public.modelStats",
-  "space.byId",
-  "space.latest",
-  "profile.byId",
-  "profile.byAddress",
-  "profile.byHandle",
-];
