+++ apps/client/src/editor/components/EditorNavbar/EditorNavbar.tsx
@@ -3,9 +3,9 @@ import { useState } from "react";
 import { MdArrowBackIosNew } from "react-icons/md";
 import { useSigner } from "wagmi";
 
+import SignInButton from "../../../../app/(navbar)/SignInButton";
 import { useEditorStore } from "../../../../app/editor/[id]/store";
 import { useSession } from "../../../client/auth/useSession";
-import SignInButton from "../../../home/SignInButton";
 import { Project } from "../../../server/helpers/fetchProject";
 import Button from "../../../ui/Button";
 import DialogContent, { DialogRoot } from "../../../ui/Dialog";
