+++ apps/client/src/play/utils/getModelStats.ts
