+++ apps/client/src/server/helpers/fetchProfileOwner.ts
@@ -1,7 +1,7 @@
 import { Profile__factory, PROFILE_ADDRESS } from "contracts";
 import { cache } from "react";
 
-import { ethersProvider } from "../constants";
+import { ethersProvider } from "../ethers";
 
 export const fetchProfileOwner = cache(async (profileId: number) => {
   const contract = Profile__factory.connect(PROFILE_ADDRESS, ethersProvider);
