+++ apps/client/src/server/helpers/fetchSpaceOwner.ts
@@ -1,7 +1,7 @@
 import { Space__factory, SPACE_ADDRESS } from "contracts";
 import { cache } from "react";
 
-import { ethersProvider } from "../constants";
+import { ethersProvider } from "../ethers";
 
 export const fetchSpaceOwner = cache(async (spaceId: number) => {
   const contract = Space__factory.connect(SPACE_ADDRESS, ethersProvider);
