+++ apps/client/src/ui/Avatar.tsx
